+++ jmix-search/search/src/main/java/io/jmix/search/index/fileparsing/FileParserResolver.java
@@ -14,26 +14,15 @@
  * limitations under the License.
  */
 
-package io.jmix.search.utils.parserresolving.resolvers;
+package io.jmix.search.index.fileparsing;
 
-import io.jmix.search.utils.parserresolving.FileParserResolver;
 import org.apache.tika.parser.Parser;
-import org.apache.tika.parser.rtf.RTFParser;
-import org.springframework.core.annotation.Order;
-import org.springframework.stereotype.Component;
 
 import java.util.List;
 
-@Component("search_PDFParserResolver")
-@Order(100)
-public class PDFParserResolver implements FileParserResolver {
-    @Override
-    public List<String> getExtension() {
-        return List.of("rtf");
-    }
+public interface FileParserResolver {
 
-    @Override
-    public Parser getParser() {
-        return new RTFParser();
-    }
+    List<String> getExtension();
+
+    Parser getParser();
 }
