+++ jmix-search/search/src/main/java/io/jmix/search/index/fileparsing/resolvers/OpenOfficeDocumentsParserResolver.java
@@ -14,26 +14,26 @@
  * limitations under the License.
  */
 
-package io.jmix.search.utils.parserresolving.resolvers;
+package io.jmix.search.index.fileparsing.resolvers;
 
-import io.jmix.search.utils.parserresolving.FileParserResolver;
+import io.jmix.search.index.fileparsing.FileParserResolver;
 import org.apache.tika.parser.Parser;
-import org.apache.tika.parser.microsoft.OfficeParser;
+import org.apache.tika.parser.odf.OpenDocumentParser;
 import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
 
-@Component("search_OldOfficeDocumentsParserResolver")
+@Component("search_OpenOfficeDocumentsParserResolver")
 @Order(100)
-public class OldOfficeDocumentsParserResolver implements FileParserResolver {
+public class OpenOfficeDocumentsParserResolver implements FileParserResolver {
     @Override
     public List<String> getExtension() {
-        return List.of("doc", "xls");
+        return List.of("odt", "ods");
     }
 
     @Override
     public Parser getParser() {
-        return new OfficeParser();
+        return new OpenDocumentParser();
     }
 }
